Rename and add packages; Reimplement methods in Configuration with reflections
