+++ src/main/java/edu/kit/aquaplanning/grounding/datastructures/ArgumentNode.java
