+++ src/main/java/edu/kit/aquaplanning/grounding/datastructures/AtomTable.java
