+++ src/main/java/edu/kit/aquaplanning/grounding/datastructures/LiftedState.java
