+++ src/main/java/edu/kit/aquaplanning/grounding/datastructures/OperatorIndex.java
