+++ src/main/java/edu/kit/aquaplanning/planning/GreedyBestFirstSearchPlanner.java
