+++ src/main/java/edu/kit/aquaplanning/planning/datastructures/SearchNode.java
