+++ src/main/java/edu/kit/aquaplanning/planning/datastructures/SearchQueue.java
