+++ src/main/java/edu/kit/aquaplanning/planning/datastructures/SearchStrategy.java
