+++ src/main/java/edu/kit/aquaplanning/planning/sat/FixedLengthSatPlanner.java
