+++ src/main/java/edu/kit/aquaplanning/planning/sat/HegemannsSatPlanner.java
