+++ src/main/java/edu/kit/aquaplanning/planning/sat/SimpleSatPlanner.java
