+++ src/main/java/edu/kit/aquaplanning/planning/sat/SymbolicReachabilityPlanner.java
